Fix bug when there are no chunks

+++ classes/storage/StorageCloudS3.class.php
@@ -241,23 +241,25 @@ class StorageCloudS3 extends StorageFilesystem
                     ];
                 }
 
-		// Create batches of requested batchsize
-                $chunked_queue = array_chunk($delete_queue, $bulk_size);
-                Logger::debug('bulk deleteFile has ' . count($chunked_queue) . ' requests');
+		if( $delete_queue ) {
+
+		    // Create batches of requested batchsize
+                    $chunked_queue = array_chunk($delete_queue, $bulk_size);
+                    Logger::debug('bulk has ' . count($chunked_queue) . ' requests');
 
 
-                // Perform actual chunk removal
-                foreach($chunked_queue as $delete_batch) {
+                    // Perform actual chunk removal
+                    foreach($chunked_queue as $delete_batch) {
 
-                    Logger::debug('bulk deleteFile deleting ' . count($delete_batch) . ' chunks');
-                    $result = $client->deleteObjects([
-                       'Bucket' => $bucket_name,
-                       'Delete' => [
-                           'Objects' => $delete_batch,
-                       ],
-                    ]);
-                }
-
+                        Logger::debug('bulk deleting ' . count($delete_batch) . ' chunks');
+                        $result = $client->deleteObjects([
+                           'Bucket' => $bucket_name,
+                           'Delete' => [
+                               'Objects' => $delete_batch,
+                           ],
+                        ]);
+                    }
+		}
 
             } else {
